refactor parsers and types
